requirements.txt
botbuilder-integration-aiohttp>=4.14.0
pandas
dotenv
requests
tabulate
azure-identity
azure-keyvault
